taint/test/mochitest/test_iframe_sinks.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test Cookie Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let string_content = "hello";
      let sink_names = [
          "iframe.src", 
          "iframe.srcdoc",
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          SimpleTest.is(report.detail.str, "tainted=hello", "Check sink string content");

          let flow = report.detail.str.taint[0].flow;
          SimpleTest.is(flow[2].operation, sink_names[i]);

	  i += 1;
	  if (i >= sink_names.length) {
              SimpleTest.finish();
	  }
      }, false);

      function startTest() {
          const frame = document.getElementById("iframe");
          const tainted_str = String.tainted("tainted=hello");

          frame.src = tainted_str;
          frame.srcdoc = tainted_str;
      }
      
    </script>
  </head>

  <body onload="startTest();">
    <iframe id="iframe"></iframe>
  </body>

</html>
